Split MuJoCo system interface into mujoco_hardware_interface package
Demote ai_sapiens_mujoco to a plugin-free simulation library and host
MujocoSystem and GantryServiceNode in a new package under
ai_sapiens_hardware_interfaces, following the radiomaster_usb layout.
