+++ ai_sapiens/package.xml
@@ -18,6 +18,7 @@
   <exec_depend>ai_sapiens_mujoco</exec_depend>
   <exec_depend>ai_sapiens_rc_broadcaster</exec_depend>
   <exec_depend>ai_sapiens_sim2real</exec_depend>
+  <exec_depend>mujoco_hardware_interface</exec_depend>
   <exec_depend>radiomaster_usb_hardware_interface</exec_depend>
 
   <export>
