+++ ai_sapiens_hardware_interfaces/mujoco_hardware_interface/CMakeLists.txt
@@ -1,0 +1,69 @@
+cmake_minimum_required(VERSION 3.16)
+project(mujoco_hardware_interface)
+
+if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
+  add_compile_options(-Wall -Wextra -Wpedantic)
+endif()
+
+find_package(ament_cmake REQUIRED)
+find_package(hardware_interface REQUIRED)
+find_package(pluginlib REQUIRED)
+find_package(rclcpp REQUIRED)
+find_package(rclcpp_lifecycle REQUIRED)
+find_package(ai_sapiens_mujoco REQUIRED)
+find_package(ai_sapiens_interfaces REQUIRED)
+find_package(std_srvs REQUIRED)
+
+add_library(${PROJECT_NAME} SHARED
+  src/gantry_service_node.cpp
+  src/mujoco_hardware_interface.cpp
+)
+target_compile_features(${PROJECT_NAME} PUBLIC cxx_std_17)
+target_include_directories(${PROJECT_NAME} PUBLIC
+  $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+  $<INSTALL_INTERFACE:include>
+)
+ament_target_dependencies(${PROJECT_NAME}
+  hardware_interface
+  pluginlib
+  rclcpp
+  rclcpp_lifecycle
+  ai_sapiens_mujoco
+  ai_sapiens_interfaces
+  std_srvs
+)
+
+pluginlib_export_plugin_description_file(
+  hardware_interface mujoco_hardware_interface.xml)
+
+install(TARGETS ${PROJECT_NAME}
+  EXPORT export_${PROJECT_NAME}
+  ARCHIVE DESTINATION lib
+  LIBRARY DESTINATION lib
+  RUNTIME DESTINATION bin
+)
+install(DIRECTORY include/ DESTINATION include)
+
+if(BUILD_TESTING)
+  find_package(ament_cmake_gtest REQUIRED)
+  find_package(ament_index_cpp REQUIRED)
+
+  ament_add_gtest(test_mujoco_system test/test_mujoco_system.cpp)
+  target_link_libraries(test_mujoco_system ${PROJECT_NAME})
+  ament_target_dependencies(test_mujoco_system
+    ament_index_cpp hardware_interface rclcpp std_srvs ai_sapiens_interfaces)
+endif()
+
+ament_export_targets(export_${PROJECT_NAME} HAS_LIBRARY_TARGET)
+ament_export_include_directories(include)
+ament_export_libraries(${PROJECT_NAME})
+ament_export_dependencies(
+  hardware_interface
+  pluginlib
+  rclcpp
+  rclcpp_lifecycle
+  ai_sapiens_mujoco
+  ai_sapiens_interfaces
+  std_srvs
+)
+ament_package()
