+++ ai_sapiens_hardware_interfaces/mujoco_hardware_interface/package.xml
@@ -1,20 +1,27 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>ai_sapiens_mujoco</name>
+  <name>mujoco_hardware_interface</name>
   <version>0.1.1</version>
-  <description>MuJoCo sim2sim simulation library for AI Sapiens robots</description>
+  <description>MuJoCo sim2sim system interface for AI Sapiens</description>
   <maintainer email="pyo@robotis.com">Pyo</maintainer>
   <license>Apache-2.0</license>
   <author email="pkw@robotis.com">Kiwoong Park</author>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
 
+  <depend>hardware_interface</depend>
+  <depend>pluginlib</depend>
   <depend>rclcpp</depend>
+  <depend>rclcpp_lifecycle</depend>
+  <depend>ai_sapiens_mujoco</depend>
+  <depend>ai_sapiens_interfaces</depend>
+  <depend>std_srvs</depend>
 
   <test_depend>ament_cmake_gtest</test_depend>
   <test_depend>ament_index_cpp</test_depend>
   <test_depend>ai_sapiens_description</test_depend>
+  <test_depend>xacro</test_depend>
 
   <export>
     <build_type>ament_cmake</build_type>
